Split up the initial 'cli_app.rs' file
All the enums (Type, Format etc.) and utility functions are now in
'util.rs'. The struct 'CliApp' is now the only thing in 'cli_app.rs'.
The handle_matches function is now in 'handle.rs'.
