+++ src/cli/mod.rs
@@ -1,5 +1,8 @@
 mod clap;
 mod cli_app;
+mod handle;
+mod util;
 
 pub use self::clap::{play_subcommand, playback_subcommand, query_subcommand};
-pub use cli_app::handle_matches;
+use cli_app::CliApp;
+pub use handle::handle_matches;
